% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigaR.R
\name{degFreedom}
\alias{degFreedom}
\title{Identify the number of degree of freedom used in the model testing}
\usage{
degFreedom(fit, design, ZSpline)
}
\arguments{
\item{fit}{Fitted object of the full model,}

\item{design}{Design matrix for replicates.}

\item{ZSpline}{Design matrix for thin-plate splines.}
}
\value{
This function return number of degree of freedom used in the model testing.
}
\description{
This function calculate the actual degree of freedom employed in the model testing.
}
