% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigaR.R
\name{getDesign}
\alias{getDesign}
\title{Design matrix corresponding to the data}
\usage{
getDesign(timefac, groupfac, numKnots, deg, diffSpl)
}
\arguments{
\item{timefac}{Vector representing the time points of the design matrix.}

\item{groupfac}{Factor representing the groups of the design matrix.}

\item{numKnots}{Integer. The optimal number of knots to be used in the fitting of the splines.}

\item{deg}{Number of the degree of freedom used in the model.}

\item{diffSpl}{Boolean. If TRUE, fit the different spline per group.}
}
\value{
This function return number of degree of freedom used in the model testing.
}
\description{
Generate design matrices corresponding to the experiment and thin-plate splines.
}
