% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigaR.R
\name{numKnots}
\alias{numKnots}
\title{Estimate optimal number of knots required for fitting mixed model with thin-plate splines}
\usage{
numKnots(
  form,
  dat,
  dat1 = NULL,
  fams,
  timefac,
  groupfac,
  deg,
  shrinksimul,
  diffSpl,
  orthogonal = FALSE,
  shrink = FALSE,
  multivar = FALSE,
  rho = 0
)
}
\arguments{
\item{form}{Formula. See INLA and f for specification of the model formula.}

\item{dat}{Matrix or data frame containing the data. Rows are features, columns are samples.}

\item{dat1}{Additional matrix or data frame containing the data. Rows are features, columns are samples.}

\item{fams}{Character string. Either equal to "poisson", "zip" (zero-inflated Poisson), "nb" (negative binomial), or "zinb" (zero-inflated negative binomial): likelihood to be used.}

\item{timefac}{Vector representing the time points of the design matrix.}

\item{groupfac}{Factor representing the groups of the design matrix.}

\item{deg}{Number of the degree of freedom used in the model.}

\item{shrinksimul}{A list object resulting from tigaRshrinkSeq or tigaRshrinkGauss.}

\item{diffSpl}{Boolean. If TRUE, fit the different spline per group.}

\item{orthogonal}{Boolean. If TRUE, ortogonalized predictor data set (DNA copy number) to the respons data set (mRNA or miRNA gene expression).}

\item{shrink}{Boolean. If TRUE, perform the shrinkage of DNA copy number neighborhoods triplets.}

\item{multivar}{Boolean. If TRUE, perform multivariate estimation of the parameters.}

\item{rho}{Number that representing the level of the correlation between the DNA copy number neighborhoods triplets.}
}
\value{
This function return number of degree of freedom used in the model testing.
}
\description{
This function estimates the optimal number of knots, required to fit the model using thin-plate splines, getting the best trade-off between the number of the parameters and the fit.
}
