% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigaR.R
\name{plot_tigaRfit}
\alias{plot_tigaRfit}
\title{Plot of the tigaR models fit}
\usage{
plot_tigaRfit(
  fit,
  fit1,
  timefac,
  groupfac,
  lattice = FALSE,
  cycle = FALSE,
  title = "title",
  multivar = FALSE
)
}
\arguments{
\item{fit}{Alternative model fit object, containing fit for all the genes.}

\item{fit1}{Null model fit object, with fit of all the genes.}

\item{timefac}{Vector representing the time points of the design matrix.}

\item{groupfac}{Factor representing the groups of the design matrix.}

\item{lattice}{Boolean. If TRUE, implement the lattice type of the plot.}

\item{cycle}{Boolean. If TRUE, fit the cycling model.}

\item{multivar}{Boolean. If TRUE, perform multivariate estimation of the parameters.}
}
\value{
This function return number of degree of freedom used in the model testing.
}
\description{
This function plot the fit of the both alternative and null tigaR models over time in order to show the variation over time of the expression levels or effect of DNA copy number.
}
