% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigaR.R
\name{tdge}
\alias{tdge}
\title{Function for temporal differential expression testing}
\usage{
tdge(dat, dat1 = NULL, fitAlt, fitNull, design, ZSpline, multivar = FALSE)
}
\arguments{
\item{dat}{Matrix or data frame containing the data. Rows are features, columns are samples.}

\item{dat1}{Additional matrix or data frame containing the data. Rows are features, columns are samples.}

\item{design}{Design matrix for replicates.}

\item{ZSpline}{Design matrix for thin-plate splines.}

\item{multivar}{Boolean. If TRUE, perform multivariate estimation of the parameters.}

\item{fit}{Alternative model fit object, containing fit for all the genes.}

\item{fit1}{Null model fit object, with fit of all the genes.}
}
\description{
This function perform temporal differential expression analysis using likelihood ratio test for comparing null and alternative model in order to determine genes with significant variation over time.
}
