% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigaR.R
\name{tigaRshrinkGauss}
\alias{tigaRshrinkGauss}
\title{Simultaneous shrinkage by empirically estimating multiple priors, Gaussian likelihoods}
\usage{
tigaRshrinkGauss(
  form,
  dat,
  dat1 = NULL,
  timefac,
  groupfac,
  ZSpline,
  maxiter = 10,
  shrinkfixed = NULL,
  shrinkaddfixed = NULL,
  shrinkrandom = NULL,
  shrinkaddrandom = NULL,
  shrinksigma = TRUE,
  mixtrand = FALSE,
  excludefornull = NULL,
  fixedmeanzero = FALSE,
  addfixedmeanzero = TRUE,
  ntag = ifelse(is.null(excludefornull), c(100, 200, 500, 1000), c(1000)),
  fixed = c(0, 1/10),
  addfixed = c(0, 1/10),
  randomprec = c(1, 10^(-5)),
  addrandomprec = c(1, 10^(-5)),
  precerr = c(1, 10^(-5)),
  diracprob0 = ifelse((mixtrand | !is.null(excludefornull)), 0.8, 0.2),
  fixedseed = TRUE,
  ndraw = 10000,
  safemode = TRUE,
  tol = ifelse((mixtrand | !is.null(excludefornull)), 0.005, 0.01),
  tolrand = 0.02,
  mliktol = 0.1,
  orthogonal = FALSE,
  shrink = FALSE,
  multivar = FALSE,
  rho = 0,
  ...
)
}
\arguments{
\item{dat}{Matrix or data frame containing the data. Rows are features, columns are samples.}

\item{dat1}{Matrix of data frame containing explanatory data set. Rows are features, columns are samples.}

\item{timefac}{Vector representing the time points of the design matrix}

\item{groupfac}{Factor representing the groups of the design matrix}

\item{maxiter}{Integer. The maximum number of iterations in the iterative marginal procedure}

\item{shrinkfixed}{String. Name of the fixed parameter for which a Gaussian prior is fit.}

\item{shrinkaddfixed}{String or character vector. Name(s) of additional fixed parameter(s) for which Gaussians priors are fit.}

\item{shrinkrandom}{String. Name of Gaussian random effects parameter for which an inverse-Gamma prior is fit to the variance.}

\item{shrinkaddrandom}{String or character vector. Name(s) of additional random effects parameter(s) for which inverse-Gamma priors are fit to the variances.}

\item{shrinksigma}{Boolean. Fit a prior to the error variance?}

\item{mixtrand}{Boolean. Add a point mass on zero to the inverse-Gamma prior for the random effect?}

\item{excludefornull}{String or character vector. Name(s) of parameter(s) to be excluded from the given model in form resulting in the null mode.}

\item{fixedmeanzero}{Boolean. Fix the mean of the Gaussian prior of the shrinkfixed parameter to zero?}

\item{addfixedmeanzero}{Boolean or list of booleans. Fix the mean(s) of the Gaussian prior of the shrinkaddfixed parameter(s) to zero?}

\item{ntag}{Integer vector. Consecutive number of data rows used for fitting the priors. Rows are samples at random with a fixed seed.}

\item{fixed}{Numeric vector of length 2. Mean and precision of initial Gaussian prior of shrinkfixed.}

\item{addfixed}{Numeric vector of length 2 or list of vectors of size 2. Mean(s) and precision(s) of initial Gaussian prior(s) of shrinkaddfixed}

\item{randomprec}{Numeric vector of length 2. Shape and rate of initial inverse-Gamma prior of shrinkrandom.}

\item{addrandomprec}{Numeric vector of length 2. Shape and rate of initial inverse-Gamma priors of shrinkaddrandom.}

\item{precerr}{Numeric vector of length 2. Shape and rate of initial inverse-Gamma prior of the error variance.}

\item{diracprob0}{Numeric between 0 and 1. Initial mixture proportion for zero-component of the prior for shrinkrandom. Only relevant when mixtrand=TRUE}

\item{fixedseed}{Boolean. Use a fixed seed for sampling the features?}

\item{ndraw}{Integer. The number of random draws from the empirical mixture of posteriors}

\item{safemode}{Boolean. If TRUE, only use features for which both models provide a fit, in case of a mixture prior.}

\item{tol}{Numeric. Tolerance for the Kolmogorov-Smirnov distance between two succesive iterations.}

\item{tolrand}{Numeric. Tolerance for the Kolmogorov-Smirnov distance between two succesive iterations for the prior of the random effect variance.}

\item{mliktol}{Numeric. Relative tolerance in percentage between two succesive iterations for the total marginal likelihood.}

\item{orthogonal}{Boolean. If TRUE, ortogonalized predictor data set (DNA copy number) to the respons data set (mRNA or miRNA gene expression)}

\item{shrink}{Boolean. If TRUE, perform the shrinkage of DNA copy number neighborhoods triplets.}

\item{multivar}{Boolean. If TRUE, perform multivariate estimation of the parameters.}

\item{rho}{Number representing the level of the correlation between the triplets.}

\item{...}{Further arguments passed on to inla.}

\item{forms}{Formula, or list of formulas the length of which equals the number of data rows. See inla and f for specification of the model formula.}

\item{Zspline}{Design matrix for thin-plate splines}
}
\value{
This function return number of degree of freedom used in the model testing.
}
\description{
This function implements the iterative joint procedure as described in Van de Wiel et al. (2012) and adapted to Gaussian likelihoods in Van de Wiel et al. (2013). It recursively applies INLA, and fits parameters of multiple priors by an EM-type algorithm.
}
